+++ lib/gravity_sensor.dart
@@ -1,7 +1,9 @@
-import 'gravity_sensor_platform_interface.dart';
+import 'src/gravity_sensor_platform_interface.dart';
+import 'src/models/gravity_event.dart';
 
-export 'src/gravity_event.dart';
+export 'src/models/gravity_event.dart';
 
+/// API for accessing information from gravity sensor of the device.
 class GravitySensor extends GravitySensorPlatform {
   GravitySensor._();