+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- **DOCS**: Provided documentation for `GravitySensor` and `GravitySensorPlatform`.
+- **CHORE**: Updated `flutter_lints` to `^5.0.0`.
+- **FEAT**: gravityEvents now can throw `FormatException` if the event format is invalid.
+
 ## 1.0.0
 
 * Initial release.